+++ src/components/Comment/styles.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+import { Card, CardHeader } from "@material-ui/core";
+
+export const CardPost = styled(Card)`
+  margin: 1rem;
+`
+
+export const CommentHeader = styled(CardHeader)`
+  text-align: center;
+  padding-bottom: 0;
+`
+
