+++ src/components/CreatePost/styles.js
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+import { Button, Typography, Paper } from "@material-ui/core";
+
+
+
+export const BoxPostWrapper = styled(Paper)`
+  margin: 1rem;
+  padding: 0 1rem;
+`
+
+export const FormCreatePost = styled.form`
+  display:flex;
+  flex-direction: column;
+`
+
+export const ButtonStyled = styled(Button)`
+  margin: 1rem auto;
+`
+
+export const TitleCreatePost = styled(Typography)`
+  text-align:center;
+  margin: 1rem;
+`
